requirements.txt
asgiref==3.7.2
Django==5.0.2
django-rest-framework==0.1.0
djangorestframework==3.14.0


